% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{branch_and_bound_rcpp}
\alias{branch_and_bound_rcpp}
\title{Affectation par méthode branch and bound}
\usage{
branch_and_bound_rcpp(G, P, R)
}
\arguments{
\item{G}{contient les groupes d'élèves}

\item{P}{contient les projets et leurs capacités maximales}

\item{R}{matrice qui contient l'ordre de préférence de chaque groupe}
}
\value{
une attribution d'élèves à des projets optimisée
}
\description{
fonction de simulation d'une attribution avec la méthode du branch and bound
}
