% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{comptage_eleve}
\alias{comptage_eleve}
\title{comptage élève algorithme using C++}
\usage{
comptage_eleve(groupe)
}
\arguments{
\item{G}{une liste de groupe d'étudiant}
}
\value{
le nombre d'étudiant total
}
\description{
comptage élève algorithme using C++
}
