% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumeration.R
\name{enumeration}
\alias{enumeration}
\title{Affectation par énumération des possibilités}
\usage{
enumeration(G, P, R)
}
\arguments{
\item{G}{une liste des groupes d'élèves}

\item{P}{une liste de projets}

\item{R}{une mactrice de voeux de chaque groupe de G}
}
\value{
la matrice d'attribution des projets par élève et par groupe
}
\description{
Affectation des élèves à un projet par ordre des voeux et d'inscription
}
