% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_glouton.r
\name{glouton_algorithm}
\alias{glouton_algorithm}
\title{Affectation par méthode gloutonne}
\usage{
glouton_algorithm(G, P, R)
}
\arguments{
\item{G}{une liste des groupes d'élèves}

\item{P}{une liste de projets}

\item{R}{une matrice de voeux de chaque groupe de G}
}
\value{
la matrice d'attribution des projets par élève et par groupe
}
\description{
Affectation des élèves à un projet par ordre des voeux et d'inscription
}
