% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_element}
\alias{is_element}
\title{is_element algorithme using C++}
\usage{
is_element(n, v)
}
\arguments{
\item{n}{un nombre}

\item{v}{un vector}
}
\value{
renvoie si n appartient à v ou non
}
\description{
is_element algorithme using C++
}
