% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotterie.R
\name{lotterie}
\alias{lotterie}
\title{Affectation par lotterie}
\usage{
lotterie(P, p, G, R)
}
\arguments{
\item{P}{une liste de projets}

\item{p}{contient les capacités maximum des projets}

\item{G}{une liste des groupes d'élèves}

\item{R}{une mactrice de voeux de chaque groupe de G}
}
\value{
la matrice d'attribution des projets par élève et par groupe
}
\description{
Affectation des élèves à un projet par ordre des voeux et d''inscription'un système de lotterie
}
