% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lottery_rcpp}
\alias{lottery_rcpp}
\title{Méthode lottery algorithm using C++}
\usage{
lottery_rcpp(P, G, R)
}
\arguments{
\item{P}{une liste de projet avec des équipes de projet}

\item{G}{une liste de groupe d'étudiant}

\item{R}{une liste de voeu classée par ordre de priorité pour chaque groupe}
}
\value{
renvoie une matrice X de taille (nb eleves,4) telle que chaque colonne correspond à l'élève, son groupe, son projet et son équipe-projet
}
\description{
Méthode lottery algorithm using C++
}
