\name{projet-package}
\alias{projet-package}
\alias{projet}
\docType{package}
\title{
  Attribution eleves-projets
}
\description{
  Ce package propose des methodes (naives et optimisees pour attribuer une liste d'equipes d'eleves avec des projets selon la preference des eleves)
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Your Name, email optional.

Maintainer: Your Name <your@email.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
