% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{simu}
\alias{simu}
\title{Simulation de l'échantillon}
\usage{
simu(nb_eleve, nb_proj, max_eleve = NULL)
}
\arguments{
\item{nb_eleve}{nombre d'élèves devant choisir un projet}

\item{nb_proj}{nombre de projets total}

\item{max_eleve}{permet de fixer le nombre maximal d'élèves dans un groupe}
}
\value{
une liste contenant : la liste des projets, la liste des groupes d'élèves et la matrice de leurs voeux respectifs
}
\description{
fonction de simulation d'un échantillon d'élèves par groupe et de projets
}
