% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{which_is_equal}
\alias{which_is_equal}
\title{Which algorithme using C++}
\usage{
which_is_equal(v, nombre)
}
\arguments{
\item{v}{un vector}

\item{nombre}{un nombre}
}
\value{
renvoie l'indice du vecteur où v[ind] = nombre (sinon renvoie -1)
}
\description{
Which algorithme using C++
}
